requirements.txt
requests
pytest
pyyaml
allure-pytest
pytest-html
loguru